move SensitiveDetector base class into SDs library
