+++ SimCore/include/SimCore/G4User/TrackingAction.h
@@ -13,7 +13,7 @@
 #include <vector>
 
 // LDMX
-#include "SimCore/TrackMap.h"
+#include "SimCore/G4User/TrackMap.h"
 
 // Geant4
 #include "G4RunManager.hh"
