+++ SimCore/include/SimCore/SDs/SensitiveDetector.h
@@ -6,7 +6,7 @@
 #include "SimCore/ConditionsInterface.h"
 #include "SimCore/Factory.h"
 #include "SimCore/G4User/TrackingAction.h"
-#include "SimCore/TrackMap.h"
+#include "SimCore/G4User/TrackMap.h"
 
 //------------//
 //   Geant4   //
