+++ SimCore/src/SimCore/G4User/EventAction.cxx
@@ -10,14 +10,12 @@
 /*   SimCore   */
 /*~~~~~~~~~~~~~*/
 #include "SimCore/G4User/TrackingAction.h"
-#include "SimCore/RunManager.h"
-#include "SimCore/TrackMap.h"
+#include "SimCore/G4User/TrackMap.h"
 
 /*~~~~~~~~~~~~*/
 /*   Geant4   */
 /*~~~~~~~~~~~~*/
 #include "G4Event.hh"
-#include "G4RunManager.hh"
 
 namespace simcore {
 namespace g4user {
