+++ SimCore/src/SimCore/Simulator.cxx
@@ -24,7 +24,7 @@
 #include "SimCore/G4User/TrackingAction.h"
 #include "SimCore/Geo/ParserFactory.h"
 #include "SimCore/PrimaryGenerator.h"
-#include "SimCore/SensitiveDetector.h"
+#include "SimCore/SDs/SensitiveDetector.h"
 #include "SimCore/UserEventInformation.h"
 #include "SimCore/XsecBiasingOperator.h"
